+++ pkg/gatewayserver/gatewayserver_util_test.go
@@ -21,6 +21,7 @@ import (
 	"net"
 	"time"
 
+	pbtypes "github.com/gogo/protobuf/types"
 	"go.thethings.network/lorawan-stack/v3/pkg/component"
 	"go.thethings.network/lorawan-stack/v3/pkg/crypto"
 	"go.thethings.network/lorawan-stack/v3/pkg/encoding/lorawan"
@@ -89,7 +90,8 @@ func (is *mockIS) add(ctx context.Context, ids ttnpb.GatewayIdentifiers, key str
 }
 
 var (
-	errNotFound = errors.DefineNotFound("not_found", "not found")
+	errNotFound        = errors.DefineNotFound("not_found", "not found")
+	errNoGatewayRights = errors.DefinePermissionDenied("no_gateway_rights", "no gateway rights")
 )
 
 func (is *mockIS) Get(ctx context.Context, req *ttnpb.GetGatewayRequest) (*ttnpb.Gateway, error) {
@@ -97,6 +99,9 @@ func (is *mockIS) Get(ctx context.Context, req *ttnpb.GetGatewayRequest) (*ttnpb
 	gtw, ok := is.gateways[uid]
 	if !ok {
 		return nil, errNotFound.New()
+	}
+	if gtw == nil {
+		return nil, errNoGatewayRights.New() // This simulates the behaviour of the IS with a deleted gateway.
 	}
 	return gtw, nil
 }
@@ -109,6 +114,16 @@ func (is *mockIS) Update(ctx context.Context, req *ttnpb.UpdateGatewayRequest) (
 	}
 	gtw.SetFields(req.Gateway, req.FieldMask.GetPaths()...)
 	return gtw, nil
+}
+
+func (is *mockIS) Delete(ctx context.Context, ids *ttnpb.GatewayIdentifiers) (*pbtypes.Empty, error) {
+	uid := unique.ID(ctx, ids)
+	_, ok := is.gateways[uid]
+	if !ok {
+		return nil, errNotFound.New()
+	}
+	is.gateways[uid] = nil
+	return nil, nil
 }
 
 func (is *mockIS) GetIdentifiersForEUI(ctx context.Context, req *ttnpb.GetGatewayIdentifiersForEUIRequest) (*ttnpb.GatewayIdentifiers, error) {